% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_indirect.R
\name{all_indirect}
\alias{all_indirect}
\title{test}
\usage{
all_indirect(data, model)
}
\arguments{
\item{data}{ggafadfb}

\item{model}{test}
}
\value{
All indirect effets in a data from the first independent variable to the last outcome.
}
\description{
test test
}
\examples{
all_indirect(model = y ~ m1 ~ x, data = medEX)
}
