% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2cov.R
\name{beta2cov}
\alias{beta2cov}
\title{Compute a covariance matrix from regression coefficients and variances}
\usage{
beta2cov(B, V = NULL)
}
\arguments{
\item{B}{A lower triangular matrix containing regression coefficients (direct effects).}

\item{V}{A vector containing the variance of the covariance matrix. If \code{V = NULL} (default), then \code{diag(S) = diag(p)} where \code{p} is the number of variables,}
}
\value{
A covariance matrix.
}
\description{
Compute a covariance matrix, \code{S}, from a lower triangular matrix containing regression coefficients (direct effects), \code{B} and the variances of the variables \code{V}.
}
\examples{
beta <- matrix(c( 0,  0,  0,
                 .4,  0,  0,
                 .2, .5,  0),
                 ncol = 3, nrow = 3, byrow = TRUE)
beta2cov(beta)
}
