% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BK.R
\name{BK}
\alias{BK}
\title{The causal step methods more widely known as the Baron-Kenny method}
\usage{
BK(out, alpha = out$alpha)
}
\arguments{
\item{out}{An output of class \code{PAmediaiton} form pathanalysis::mediation.}

\item{alpha}{Type I error rate. Default is the one from the original mediation.}
}
\value{
Determine if there is a mediated relation for each indirect effects and whether it is partial or complete.
}
\description{
Perform the Baron-Kenny method on a recursive path model.
}
\examples{
res <- mediation(model = y ~ m2 ~ m1 ~ x, data = medEX, standardized = TRUE)
BK(res)
}
