% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot}
\alias{boot}
\title{Bootstrap}
\usage{
boot(data, FUN, alpha = 0.05, nrep = 5000, ...)
}
\arguments{
\item{data}{The data sets.}

\item{FUN}{A function comuting the desired statistics.}

\item{alpha}{Type I error rate. Default is .05.}

\item{nrep}{Number of repetitions for the boostrap. Default is 5000.}

\item{...}{Optional arguments for \code{FUN}.}
}
\value{
Return the estimates computed by \code{FUN}, their standard error and their \eqn{100 * (1-\alpha)}\% confidence interval.
}
\description{
A generic function to boostrap the desired statistics.
}
\examples{
boot(data = rpois(n = 10, lambda = 5), FUN = median, nrep = 10)
}
