% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2beta.R
\name{cov2beta}
\alias{cov2beta}
\title{Compute all regression coefficients from a covariance matrix}
\usage{
cov2beta(S)
}
\arguments{
\item{S}{A covariance matrix.}
}
\value{
A lower triangular matrix containing all direct effects (regression coefficients)
}
\description{
Compute all direct regression coefficients (direct effects), \code{B} from a covariance matrix starting from the first variable to the last (order matter).
}
\examples{
S <- matrix(c( 1, .2, .4,
              .2,  1, .3,
              .4, .5,  1), ncol = 3, nrow = 3)
cov2beta(S)
}
