% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medEX.R
\docType{data}
\name{medEX}
\alias{medEX}
\title{Hypothetical data set for serial mediation analysis}
\format{
A data frame with 432 rows and 5 variables:
\describe{
\item{id}{Subjects' identifiers}
\item{x}{First independent variable}
\item{m1}{First mediatior}
\item{m2}{Second mediator}
\item{y}{Outcome variable}
}
}
\source{
\url{https://www.tqmp.org/}
}
\usage{
medEX
}
\description{
A dataset containing 4 variables generated with from a
serial mediation model
}
\keyword{datasets}
