% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{Mediation Analysis}
\usage{
mediation(
  model,
  data = NULL,
  stat = all_indirect,
  nrep = 5000,
  alpha = 0.05,
  standardized = FALSE
)
}
\arguments{
\item{model}{an object of class \code{"formula"}: a symbolic description of the model to be fitted. More details of model specification are given under "Details". Basically, \code{model} is the order of the variables in the model from last (outcome variables) to first (exogenous variables).}

\item{data}{an optional data frame, list or environment (or object coercible by \code{as.data.frame()} to a data frame) containing the variables in the model. If \code{data = NULL}, the variables are taken from environment(formula), typically the environment from which \code{mediation} is called.}

\item{stat}{A function for the statistics to compute either \code{all_indirect} (saturated model) or \code{specific_indirect}}

\item{nrep}{Number of repetitions for the boostrap. Default is 5000.}

\item{alpha}{Type I error rate to build the \eqn{100 * (1-\alpha)}\% confidence interval. Default is .05.}

\item{standardized}{Option to get standardized coefficients.}
}
\value{
The mediation analysis
}
\description{
Return all direct and indirect effects within a recursive path model.
}
\details{
Basically, \code{mediation()} is a wrapper for the \code{\link{boot()}} function from the  \code{pathanalysis} package.
}
\examples{
mediation(model = y ~ m2 ~ m1 ~ x, data = medEX, standardized = TRUE)
}
