% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PAmediation.R
\name{print.PAmediation}
\alias{print.PAmediation}
\title{Print results of mediation}
\usage{
\method{print}{PAmediation}(x, ...)
}
\arguments{
\item{x}{An object of class "PAmediation".}

\item{...}{Further arguments for other methods, ignored for "mediation".}
}
\description{
Print the results of the mediation function.
}
\examples{
\dontrun{
results <- mediation(model = y ~ m2 ~ m1 ~ x, data = medEX, standardized = TRUE)
print(results)
}
}
