% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_indirect.R
\name{specific_indirect}
\alias{specific_indirect}
\title{Return a Specific Indirect Effect}
\usage{
specific_indirect(data, model)
}
\arguments{
\item{data}{test}

\item{model}{test}
}
\value{
test
}
\description{
test test
}
\examples{
all_indirect(model = y ~ m2 ~ m1 ~ x, data = medEX)
}
